refactor: 인접한 필드를 오픈하는 책임 이동 (지뢰 게임 -> 지뢰 판 필드들)
