+++ src/main/kotlin/minesweeper/domain/BoardFields.kt
@@ -7,18 +7,25 @@ data class BoardFields(val boardFields: List<BoardField>) {
             .count { it is MineField }
     }
 
-    fun open(coordinate: Coordinate): BoardField {
+    fun open(coordinate: Coordinate) {
         val boardField = boardFields.find { it.coordinate == coordinate }
             ?: throw IllegalArgumentException("해당 좌표에 필드가 존재하지 않습니다. (${coordinate.x},${coordinate.y})")
-        boardField.open()
 
-        return boardField
+        when (boardField) {
+            is MineField -> boardField.open()
+            is NumberField -> boardField.open()
+                .also { openAdjacentNumberFields(boardField.coordinate) }
+        }
     }
 
-    fun openNumberFields(coordinates: List<Coordinate>): List<BoardField> {
-        return boardFields.filter { coordinates.contains(it.coordinate) }
+    private fun openAdjacentNumberFields(coordinate: Coordinate) {
+        val adjacentCoordinates = coordinate.adjacentCoordinates()
+        val openNumberFields = boardFields.filter { adjacentCoordinates.contains(it.coordinate) }
             .filterIsInstance<NumberField>()
+            .filterNot { it.isOpen }
             .onEach { it.open() }
+
+        openNumberFields.forEach { openAdjacentNumberFields(it.coordinate) }
     }
 
     fun isAllOpenedNumberFields(): Boolean {
