+++ src/test/kotlin/minesweeper/domain/BoardFieldsTest.kt
@@ -1,7 +1,7 @@
 package minesweeper.domain
 
-import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.DescribeSpec
+import io.kotest.matchers.collections.shouldContainAll
 import io.kotest.matchers.collections.shouldContainExactly
 import io.kotest.matchers.shouldBe
 
@@ -32,21 +32,46 @@ class BoardFieldsTest : DescribeSpec({
     }
 
     describe("open") {
-        context("필드들에 해당하는 좌표를 입력받아") {
-            it("해당하는 필드를 열 수 있다.") {
+        context("좌표가 숫자필드 일 경우") {
+            it("숫자필드와 인접한 숫자필드들도 모두 열린다") {
+                /**
+                 * * C *
+                 * C C C
+                 * * C C
+                 */
                 val boardFields = boardFields()
-                boardFields.open(Coordinate(0, 0))
+                boardFields.open(Coordinate(0, 1))
 
-                val boardField = boardFields.boardFields.first() { it.coordinate == Coordinate(0, 0) }
-                boardField.isOpen shouldBe true
+                val openedCoordinates = boardFields.boardFields
+                    .filter { it.isOpen }
+                    .map { it.coordinate }
+                openedCoordinates shouldContainAll listOf(
+                    Coordinate(0, 1),
+                    Coordinate(1, 0),
+                    Coordinate(1, 1),
+                    Coordinate(1, 2),
+                    Coordinate(2, 1),
+                    Coordinate(2, 2),
+                )
             }
         }
 
-        context("필드들에 해당하지 않는 좌표를 입력받으면") {
-            it("IllegalArgumentException 이 발생한다.") {
+        context("좌표가 지뢰필드 일 경우") {
+            it("지뢰필드는 열리지만 인접한 필드들은 열리지 않는다") {
+                /**
+                 * * C *
+                 * C C C
+                 * * C C
+                 */
                 val boardFields = boardFields()
+                boardFields.open(Coordinate(0, 0))
 
-                shouldThrow<IllegalArgumentException> { boardFields.open(Coordinate(3, 3)) }
+                val openedCoordinates = boardFields.boardFields
+                    .filter { it.isOpen }
+                    .map { it.coordinate }
+                openedCoordinates shouldContainExactly listOf(
+                    Coordinate(0, 0)
+                )
             }
         }
     }
@@ -60,7 +85,9 @@ class BoardFieldsTest : DescribeSpec({
         context("필드들에 해당하는 좌표들 입력받아") {
             it("숫자필드 인 경우 필드를 열 수 있다") {
                 val boardFields = boardFields()
-                boardFields.openNumberFields(listOf(Coordinate(1, 0)))
+                boardFields.boardFields.filter { listOf(Coordinate(1, 0)).contains(it.coordinate) }
+                    .filterIsInstance<NumberField>()
+                    .onEach { it.open() }
 
                 val openedFields = boardFields.boardFields.filter { it.isOpen }.map { it.coordinate }
                 openedFields shouldContainExactly listOf(Coordinate(1, 0))
@@ -68,7 +95,9 @@ class BoardFieldsTest : DescribeSpec({
 
             it("지뢰필드 인 경우 필드를 열지 않는다") {
                 val boardFields = boardFields()
-                boardFields.openNumberFields(listOf(Coordinate(0, 0)))
+                boardFields.boardFields.filter { listOf(Coordinate(0, 0)).contains(it.coordinate) }
+                    .filterIsInstance<NumberField>()
+                    .onEach { it.open() }
 
                 val openedFields = boardFields.boardFields.filter { it.isOpen }.map { it.coordinate }
                 openedFields shouldContainExactly emptyList()
@@ -76,7 +105,9 @@ class BoardFieldsTest : DescribeSpec({
 
             it("필드들에 해당하지 않는 필드의 경우 아무 필드도 열지 않는다") {
                 val boardFields = boardFields()
-                boardFields.openNumberFields(listOf(Coordinate(3, 3)))
+                boardFields.boardFields.filter { listOf(Coordinate(3, 3)).contains(it.coordinate) }
+                    .filterIsInstance<NumberField>()
+                    .onEach { it.open() }
 
                 val openedFields = boardFields.boardFields.filter { it.isOpen }.map { it.coordinate }
                 openedFields shouldContainExactly emptyList()
